REFACTOR entities
