+++ TheMovieApp/TheMovieApp/APIManager/APIConstants.swift
@@ -28,8 +28,8 @@ public enum MovieLanguage: String {
 // MARK: API
 public enum APIUrls: String {
   case movie = "https://api.themoviedb.org/3/movie/"
-  case tv = "https://api.themoviedb.org/3/tv/"
   case img = "http://image.tmdb.org/t/p/w500/"
+  case tv = "https://api.themoviedb.org/3/tv/"
 }
 
 public enum APIParams: String {
